Erste paar Aufgaben gelöst

+++ Challenge3/main.c
@@ -3,7 +3,9 @@
 
 int main()
 {
-    int zahl = 12.3;
+    float zahl = 12.3;
     printf("%f",zahl);
     return 0;
 }
+
+//gelöst
